refactor(metafetch): extract MetadataStateService from server
Moves MetadataStateService and its 7 tests from internal/server to
internal/metafetch. Exports MetadataFieldState (was unexported) so
server_metadata.go and handler tests can reference it. All server
callsites updated to use metafetch.MetadataStateService.
